% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{curlmemory}
\alias{curlmemory}
\title{A curl memory test function}
\usage{
curlmemory(url = "https://jsonplaceholder.typicode.com/users")
}
\description{
A curl memory test function
}
\examples{
curlmemory()
}
